preproc, register, searchkw, symbol
